requirements.txt
asgiref==3.8.1
Django==5.2
django-polymorphic==3.1.0
python-decouple==3.8
setuptools==78.1.0
sqlparse==0.5.3
